requirements.txt
crewai>=0.60.0
pydantic<3
python-dotenv
pandas
openpyxl
numpy
streamlit
duckduckgo-search>=6.2.6
tavily-python>=0.5.0
huggingface_hub>=0.23.0
llama-cpp-python[server]>=0.2.86
